Fixed Map issue
