+++ src/constants/constants.ts
@@ -1,1 +1,2 @@
 export const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL || "";
+export const DEFAULT_COORDINATES = [0,0];
